Add "payments" component (#5)
Payments are a separate and simplified form of expenses for representing a transfer to settle the debt from the expenses. The payment is included in the balance computation.

The purpose of introducing this as a distinct component is that, for instance A owes B $8 but then send them $10. The balances would then adjust to show that someone should send $2 to B.