Add support for deleting expenses and payments (#13)
The entries are deleted by clicking the trash icon in the relevant table. There is currently no confirmation flow.

Tooltip initialization was changed to use a 'MutationObserver' which does tooltip management as they're added and removed from the DOM (this also fixes the hot-reload issue).

When an entry is deleted, focus is put on the "add" button of the table.