Add persistence into local storage
The model is stored in a single field "data" as JSON prepended by a "schema" version. The data itself is also versioned by a "revision" number which is stored under a separate key "data@revision". The revision number is used to detect if the model changed in another browser window.

The update function of all components now return a "model changed" value next to the model itself. If true, then the full model is reserialized and persisted under a new revision.

This is a big change, but still, many things have been left unpolished:
- The rendering of the storage selector could be improved.
- The use case of "no storage" is unclear - it should probably be replaced by the option of having multiple instances. You would start by being in the "default" instance which you can either clear or migrate to another name (which optionally adds remove storage).
- If writes fails due to revision error, only a message is logged in the console (and the message is not too informative). In the UI, the error is silent.