Storage: Fix race condition on port subscription
When reading from or writing to storage, the JavaScript handler sends back a message with either the loaded data or the storage result (revision and error). The subscription to handle this message was conditional on the storage mode such that the message only went through if it the mode was set to "local". But when switching the mode from "none" to "local", the state is flushed to storage and the new revision wired back to the app. Apparently, this message is sent before the subscriptions are updated; i.e. we start listening for the message right after it went through.