Keep IDs of all domain types as integers
Currently the IDs of expenses and payments (not participants) are converted to (integer) strings when creating the resource to avoid doing it in the view function. But it does seem like a somewhat pointless optimization that adds just a little too much complexity (particularly when importing from stogage) to be worth it.