Disallow self payments (#21)
The check if the payer and receiver IDs are identical is done in the view functions and determines whether the "Submit" button for creating the payment should be disabled and whether a specific feedback error should be rendered.

To ensure that errors and suggested amount is rendered on the initial "create" model, it now chains the messages 'CreateEditPayer' and 'CreateEditReceiver' in after the handling of 'LoadCreate'. This change induced some refactorings to the update chaining functions. Furthermore, to avoid greeting the user with a fresh modal full of validation errors for decisions that they didn't make, the form now looks up the "first" candidates (that is, first pair of participants where an amount is actually owed). If no such candidate is found, it falls back to the first two participants. The code for this part admittedly is less than magnificent.

Part of #17.