Persist app configuration (#36)
The current use of the term "locale" has mostly been replaced with "config". The locale is considered part of the configuration and independent from the data set.

The only configuration that is currently defined is the number of decimal places and decimal separator symbol for amounts. The former is crucial for decoding the state correctly and is thus included in the persisted state. The latter is considered part of the locale which is not yet modifiable nor persisted.