Find existing payment when computing suggested payments
The payments were previously found only when applying the payment. Doing it "eagerly" this way allows the view to explicitly state what the apply command is going to do.

Some refactorings had to be done to make this feasible. Most importantly, the 'addPayments' function was inlined as it was in the way of handling the more complex logic of amending suggested payments. After other recent refactorings, it didn't serve much of a purpose any longer anyway.