Hide no-op buttons instead of removing them
The 'apply all' and 'delete all planned' buttons are currently only rendered if clicking them has any effect. Likewise, the icon for deleting a payment is hidden if the payment is marked as done. This makes the UI flicker when the buttons are added or removed. Making them invisible at least reduces this problem.