Implement "settled by" support
An optional field 'settledBy' is added to participants to indicate that the balance of the given participant should be settled by the referenced participant.

The ability to resolve the relation recursively has been implemented but is not yet used. There are probably some complexities that need to be sorted out if we are to support setting up a chain of such relations.